fix syntax table for emacs kivy-mode
"<" and ">" were considered part of symbols, but are now punctuation. This means
selecting the symbol-at-point will now correctly select "Foo" instead of "<Foo>".

"_" was part of a word, but is now part of a symbol. This means you can move by
word inside of camel_case symbols, like in other emacs modes.